+++ packages/extras/src/lib/index.ts
@@ -1,3 +1,5 @@
+import Align from './components/Align/Align.svelte'
+
 // Hooks
 export { useCursor } from './hooks/useCursor'
 export { useGltf } from './hooks/useGltf'
@@ -20,7 +22,15 @@ export { default as InstancedMesh } from './components/Instancing/InstancedMesh.
 export { default as Instance } from './components/Instancing/Instance.svelte'
 export { default as InstancedMeshes } from './components/Instancing/InstancedMeshes/InstancedMeshes.svelte'
 export { default as SoftShadows } from './components/SoftShadows/SoftShadows.svelte'
-export { default as Center } from './components/Center/Center.svelte'
+export { default as Sky } from './components/Sky/Sky.svelte'
+export { Align }
+
+/**
+ * @component Center
+ * @deprecated Please use `<Align>` instead.
+ */
+const Center = Align
+export { Center }
 
 // suspense
 export { default as Suspense } from './suspense/Suspense.svelte'