+++ packages/rapier/src/lib/components/CollisionGroups/CollisionGroups.svelte.d.ts
@@ -1,4 +1,4 @@
-import { SvelteComponentTyped } from 'svelte'
+import { SvelteComponent } from 'svelte'
 
 type N = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 | 14 | 15
 
@@ -26,6 +26,6 @@ export type CollisionGroupsProps<TGroupsDef extends GroupsDef> = TGroupsDef exte
   ? GroupsProps
   : MembershipsAndFilterProps
 
-export default class CollisionGroups<TGroupsDef extends GroupsDef> extends SvelteComponentTyped<
+export default class CollisionGroups<TGroupsDef extends GroupsDef> extends SvelteComponent<
   CollisionGroupsProps<TGroupsDef>
 > {}