+++ packages/rapier/src/lib/types/types.ts
@@ -62,11 +62,11 @@ export type ColliderEventMap = {
   }
 }
 
-export type CollisionEnterEvent = CustomEvent<ColliderEventMap['collisionenter']>
-export type CollisionExitEvent = CustomEvent<ColliderEventMap['collisionexit']>
-export type SensorEnterEvent = CustomEvent<ColliderEventMap['sensorenter']>
-export type SensorExitEvent = CustomEvent<ColliderEventMap['sensorexit']>
-export type ContactEvent = CustomEvent<ColliderEventMap['contact']>
+export type CollisionEnterEvent = ColliderEventMap['collisionenter']
+export type CollisionExitEvent = ColliderEventMap['collisionexit']
+export type SensorEnterEvent = ColliderEventMap['sensorenter']
+export type SensorExitEvent = ColliderEventMap['sensorexit']
+export type ContactEvent = ColliderEventMap['contact']
 
 export type RigidBodyEventMap = ColliderEventMap & {
   sleep: void